Migrate booleans
