+++ frontend/src/metabase-lib/constants.ts
@@ -1,0 +1,1 @@
+export const BOOLEAN_FILTER_OPERATORS = ["=", "is-null", "not-null"] as const;
